Refactoring update
Moves shot logic to its on js.file.
